requirements.txt
python-telegram-bot==20.6
python-dotenv==1.0.0
aiosqlite==0.19.0
SQLAlchemy==2.0.25
httpx